+++ core/insfile.h
@@ -80,9 +80,9 @@ struct instanceFileData
    long                           EnvBinaryLoadInstances(void *,const char *);
 #endif
    long                           EnvLoadInstances(void *,const char *);
-   long                           EnvLoadInstancesFromString(void *,const char *,int);
+   long                           EnvLoadInstancesFromString(void *,const char *,size_t);
    long                           EnvRestoreInstances(void *,const char *);
-   long                           EnvRestoreInstancesFromString(void *,const char *,int);
+   long                           EnvRestoreInstancesFromString(void *,const char *,size_t);
 
 #if ALLOW_ENVIRONMENT_GLOBALS
 
@@ -93,9 +93,9 @@ struct instanceFileData
    long                           BinarySaveInstances(const char *,int);
 #endif
    long                           LoadInstances(const char *);
-   long                           LoadInstancesFromString(const char *,int);
+   long                           LoadInstancesFromString(const char *,size_t);
    long                           RestoreInstances(const char *);
-   long                           RestoreInstancesFromString(const char *,int);
+   long                           RestoreInstancesFromString(const char *,size_t);
    long                           SaveInstances(const char *,int);
    
 #endif /* ALLOW_ENVIRONMENT_GLOBALS */
